Add /project as base slug for docs
'/' gets redirected to '/projects/backend'
